Refactor regression tests and take care of already set env variables
In environments where SRVLOG_PATH, ARCLOG_PATH and BACKUP_PATH are already
set this test suite complains with a couple of diffs that should be avoided.
In order to fix that, test variables are set after unsetting a huge set of
environment variables for clarity, adding at the same time the *_PATH
variables. The test suite is refactored at the same time to have a single
entry point to set the environment for each test. This seems more robust,
and results in quite a bit of duplicated code removal.

This patch was created by Michael Paquier and modified by me.