Improve error functions and error level definitions.
In current implementation, there are some ERROR_* and used in elog,
but these are not error levels but error cause.

This patch introduce ereport function and improve error level
definitions. The ereport can use error level and error code to be
used in exit code separately.

This patch also introduce DEBUG error level for some detail message
to be shown with debug command line option.